+++ public/manifest.json
@@ -13,12 +13,11 @@
     "48": "logo48.png",
     "128": "logo128.png"
   },
-  "manifest_version": 2,
+  "manifest_version": 3,
   "oauth2": {
     "client_id": "728982193090-p33gadj656ntoskdrp91qn99ifvg0705.apps.googleusercontent.com",
     "scopes": ["https://www.googleapis.com/auth/calendar.readonly"]
   },
   "permissions": ["geolocation", "identity"],
-  "version": "0.1.0",
-  "content_security_policy": "script-src 'self' https://accounts.google.com/gsi/client; script-src 'self' 'wasm-unsafe-eval'; object-src 'self';"
+  "version": "0.1.0"
 }
